+++ src/App.jsx
@@ -15,14 +15,18 @@ import {
 import AuthFlip from './pages/user/AuthFlip';
 import RestaurantsPage from './pages/restaurants/RestaurantsPage';
 import StorePage from './pages/store/StorePage';
-// import RestaurantDetail from './components/restaurantDetails/index';
 import { KAKAO_API_KEY } from './constants/EnvContant';
 import { ApiDocs } from './pages/api-docs/ApiDocs';
+import FavList from './components/layouts/FavList';
 
 export default function App() {
   const [isShowSideNav, setIsShowSideNav] = useState(false);
   const showSideNav = () => setIsShowSideNav(true);
   const hideSideNav = () => setIsShowSideNav(false);
+
+  const [isShowFavList, setIsShowFavList] = useState(false);
+  const showFavList = () => setIsShowFavList(true);
+  const hideFavList = () => setIsShowFavList(false);
   const { Kakao } = window;
 
   if (Kakao && !Kakao.isInitialized()) {
@@ -32,8 +36,9 @@ export default function App() {
   return (
     <>
       <Router>
-        <Header showSideNav={showSideNav} />
+        <Header showSideNav={showSideNav} showFavList={showFavList} />
         <SideNav isShowSideNav={isShowSideNav} hideSideNav={hideSideNav} />
+        <FavList isShowSideNav={isShowFavList} hideSideNav={hideFavList} />
         <Box component="main">
           <Routes>
             <Route path={PATH_USER_SIGNIN} element={<AuthFlip />} />