+++ changelog.md
@@ -26,12 +26,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 written by esther615kim
 
+`Summary`
+- Navbar 검색창 추가 & 반응형, FavList 추가 
+
  `Changed`
-- path={PATH_RESTAURANTS_DETAIL} to <StorePage />
+- path={PATH_RESTAURANTS_DETAIL} to StorePage
+- App.jsx to add FavList and status
+- Header/index.jsx to add responosive styling, rename components('Styled-element') and add menu buttons
 
 `Added`
 - cartSlicer
+- FavList component
 - Basket folder in src/components
+
 
 ## [0.1.1] - 2022-01-09
 `Setting`
@@ -58,7 +65,7 @@ written by Minju
 - 메뉴 카테고리 선택시 그 섹션으로 스크롤 이동
 
 `Changed`
-
+- changed component name from ShopDetail to RestaurantDetail
 - 모달창 x 버튼 색깔이 가려져서 변경
 
 ## [0.0.9] - 2022-01-03